Добавил подтверждение на удаление поста, исправил удаление, добавил лоудер

+++ src/components/Main/Posts/Post/Post.jsx
@@ -43,7 +43,10 @@ function Post({
   const description = text.length > 50 ? `${text.slice(0, 50)}...` : text
   const updatedDate = new Date(updated_at).toLocaleString()
   const avatarDefault = 'https://thumbs.dreamstime.com/b/%D0%B7%D0%BD%D0%B0%D1%87%D0%BE%D0%BA-%D0%BF%D0%BE-%D1%83%D0%BC%D0%BE%D0%BB%D1%87%D0%B0%D0%BD%D0%B8%D1%8E-%D0%BF%D0%BB%D0%BE%D1%81%D0%BA%D0%B8%D0%B9-%D0%B0%D0%B2%D0%B0%D1%82%D0%B0%D1%80-%D0%BF%D1%80%D0%BE%D1%84%D0%B8%D0%BB%D1%8F-%D1%81%D0%BE%D1%86%D0%B8%D0%B0%D0%BB%D1%8C%D0%BD%D1%8B%D0%B9-%D0%B2%D0%B5%D0%BA%D1%82%D0%BE%D1%80-184330869.jpg'
+  const imageDefault = 'https://cdn5.vectorstock.com/i/1000x1000/97/69/no-like-icon-sign-on-white-background-symbol-vector-26149769.jpg'
   const avatar = author ? author.avatar : avatarDefault
+  let imagePost = image
+  if (image === 'https://react-learning.ru/posts') { imagePost = imageDefault }
 
   const dispatch = useDispatch() // достаем dispatch
 
@@ -59,11 +62,14 @@ function Post({
 
   // функция удаления поста
   const deleteHandler = () => {
-    dispatch(deletePostQuery(_id))
+    const isDelete = confirm('Точно хотите удалить пост?')
+    if (isDelete) {
+      dispatch(deletePostQuery(_id))
+    }
   }
 
   // eslint-disable-next-line no-underscore-dangle
-  const isLike = likesFromRedux.includes(author._id)
+  const isLike = likesFromRedux.includes(author?._id)
 
   // поставить или удалить лайк по клику
   const likeHandler = () => {
@@ -100,7 +106,7 @@ function Post({
         <CardMedia
           component="img"
           height="200"
-          image={image}
+          image={imagePost}
           alt={title}
           sx={{
             mt: 0,
@@ -115,7 +121,7 @@ function Post({
           </Typography>
         </CardContent>
         <Box sx={{
-          mt: 'auto',
+          mb: 0,
         }}
         >
           <CardActions disableSpacing sx={{ display: 'flex', justifyContent: 'space-between' }}>
@@ -130,6 +136,7 @@ function Post({
             sx={{
               display: 'flex',
               justifyContent: 'space-between',
+              alignItems: 'center',
             }}
           >
             <Tooltip title="Лайк">
